feat(prd-admin): 团队动态改控制台三栏布局——两侧空白变统计面板
- 左栏成员统计：团队脉搏总量（滚动数字+环比）+ 成员排行 Top 10（点击下钻）
- 中栏时间线吃满剩余宽度，右栏分类统计：模块分布（占比条+点击下钻）/ 动作类型 Top 10（注册表图标+模块色）/ 活跃时段曲线
- 两侧统计栏各自独立滚动，整页上限 1840px 防巨幕失衡
- 后端 stats 端点新增动作类型分布聚合（group by Action，标签从白名单注册表解析，历史动作兜底原始 key）
- 删除单条横幅 PulseBand，组件重构为 StatsPanels（MemberStatsPanel + CategoryStatsPanel）
