+++ changelogs/2026-06-12_team-activity-pulse.md
@@ -4,3 +4,5 @@
 | polish | prd-admin | 团队脉搏产品化打磨：默认范围改「今天」、动作总量带环比趋势（较昨日/上周/上月）、模块图例与成员排行可点击下钻筛选、单模块时隐藏零信息量比例条 |
 | style | prd-admin | 团队动态视觉对标业界（GitHub/GitLab/Linear）：时间线改 rail + 头像角标动作图标 + 吸顶日期头，活跃时段改平滑面积曲线，脉搏面板加分栏线与氛围光；标题恢复全文显示，匿名模式仅隐藏成员姓名 |
 | style | prd-admin | 团队动态超宽屏排版修正：页面居中限宽 1240px 消除面板真空区与超长视线距离；视觉去「圆润饱满」——药丸统一小圆角、能量条/排行条细线化、撤掉径向柔光 |
+| feat | prd-admin | 团队动态改控制台三栏布局：左栏成员统计（脉搏总量+环比+成员排行）、中栏时间线、右栏分类统计（模块分布/动作类型/活跃时段），两侧统计可点击下钻 |
+| feat | prd-api | 团队动态 stats 端点新增动作类型分布聚合（Top 10，标签来自白名单注册表） |
